Add switch component to replace portfolio switches
