+++ src/components/Portfolio/Portfolio.js
@@ -1,30 +1,16 @@
 import './Portfolio.css';
 import Project from '../Project/Project.js';
+import Switcher from '../Switcher/Switcher.js';
 import {projectData} from '../../data/projects/proj-data.js';
 import { useState } from 'react';
 
 function Portfolio() {
-  const [project, setProject] = useState(0)
-
-  const changeProject = (e, n) => {
-    e.preventDefault();
-    if (project === projectData.length-1 && n === 2) {
-      setProject(0)
-    } else if (project === 0 && n === 0) {
-      setProject(projectData.length-1)
-    } else {
-      setProject(project + n - 1)
-    }
-  }
+  const [project, setProject] = useState(0);
 
   return(
     <section>
       <div className="title">
-        <section className="title-interface">
-          <button className="port-btn" id="back" onClick={event => changeProject(event, 0)}>«</button>
-          <h2 className="caption">Past Projects</h2>
-          <button className="port-btn" id="forward" onClick={event => changeProject(event, 2)}>»</button>
-        </section>
+        <Switcher name={'Past Projects'} item={project} setItem={setProject} gallery={projectData} />
       </div>
       <Project proj={projectData[project]}/>
     </section>
