+++ src/components/Project/Project.css
@@ -32,7 +32,7 @@
 
 .proj-abstract {
   grid-area: abst;
-  padding: 0px 10px;
+  padding: 0px 20px;
   font-size: 1.85vh;
   align-self: center;
 }
